hwtest-runner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hwtest-runner"
version = "0.1.0"
description = "Web-based test execution service for hwtest"
requires-python = ">=3.10"
license = {text = "Unlicense"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "hwtest-core",
    "hwtest-testcase",
    "hwtest-rack",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "pyyaml>=6.0",
    "httpx>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "black>=24.0",
    "pylint>=3.0",
    "types-PyYAML",
]

[project.scripts]
hwtest-runner = "hwtest_runner.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pylint.main]
py-version = "3.10"

[tool.pylint.format]
max-line-length = 100

[tool.pylint."messages control"]
disable = [
    "too-few-public-methods",
]
